// src/app/interfaces/interfaces.ts
export interface RespuestaPost {
  ok: boolean;
  pagina: number;
  post: Post[];
}

export interface Post {
  imgs?: string[];
  _id?: string;
  mensaje?: string;
  coords?: string;
  usuario?: Usuario;
  created?: string;
}

export interface Usuario {
  avatar?: string;
  _id?: string;
  nombre?: string;
  email?: string;
  password?: string
}